refactor: Split package into main and test directories
